+++ src/StyledApp.ts
@@ -5,6 +5,7 @@ import media from 'globalConfig/mobile'
 const StyledApp = styled.section`
 /* MOBILE FIRST DEFAULT STYLES */
   max-width: 100%;
+  padding-bottom: 50px;
 
   #sidebarWrap {
     grid-area: sidebar;