+++ src/components/layouts/Skills/StyledSkills.ts
@@ -1,0 +1,8 @@
+// Library Imports
+import styled from 'styled-components'
+
+const StyledSkills = styled.section`
+  padding: 20px;
+`
+
+export default StyledSkills