+++ src/components/modules/AccordionWrap/StyledAccordionWrap.ts
@@ -1,0 +1,9 @@
+// Library Imports
+import styled from 'styled-components'
+
+const StyledAccordionWrap = styled.section`
+  width: 100%;
+  margin: 2rem auto;
+`
+
+export default StyledAccordionWrap