+++ src/components/modules/HeroImage/HeroImage.tsx
@@ -1,0 +1,29 @@
+// Framework Imports
+import React from 'react'
+
+// CSS Imports
+import StyledHeroImage from './StyledHeroImage'
+
+/**
+ * This the HeroImage component.
+ */
+const HeroImage = () => {
+  return (
+    <StyledHeroImage>
+      <div>
+        <h1>Heya! I&apos;m Karl Chvojka.</h1>
+        <p>
+          I am a Full-stack developer with 10 years of industry experience building websites,
+          advertising campaigns, interacting with traditional media, and working in teams of people
+          from a variety of disciplines.
+        </p>
+        <p>
+          I have an unending curiosity for new ideas and processes that helps expand my work.
+          The advancement of technology continues to bewilder and inspire me.
+        </p>
+      </div>
+    </StyledHeroImage>
+  )
+}
+
+export default HeroImage