+++ src/globalAssets/data/resume.js
@@ -1,0 +1,187 @@
+/**
+ * Define data for the Work Experience section.
+ */
+
+const resume = {
+  workExperience: {
+    title: 'Work Experience',
+    placements: [
+      {
+        company: 'Open Mind Developments',
+        dates: 'Nov 2022 - Dec 2022',
+        desc: 'Front-End Development focusing on Customer interaction with the Shopify platform.',
+        keywords: [
+          'Front-End Development (HTML5, CSS3, CSS Grid, Flexbox)',
+          'JavaScript',
+          'jQuery',
+          'Tailwind',
+          'Shopify',
+          'ESLint',
+          'JSX',
+          'Developement Operations',
+          'Gitlab CI/CD',
+        ],
+        position: 'Front-End Developer'
+      },
+      {
+        company: 'Kepler Communications',
+        dates: 'Mar 2021 - Jul 2022',
+        desc: 'Front-End developer building interfaces for handling Communication Satellite monitoring and configuration, and configuring continuous integration, deployments and testing for each.',
+        keywords: [
+          'Front-End Development (HTML5, CSS3, CSS Grid, Flexbox)',
+          'JavaScript',
+          'jQuery',
+          'Bootstrap',
+          'ReactJS',
+          'Cypress.io',
+          'Jest',
+          'React Testing Library',
+          'ESLint',
+          'JSX',
+          'Developement Operations',
+          'Gitlab CI/CD',
+          'Semantic Ui'
+        ],
+        position: 'Web Application Developer'
+      },
+      {
+        company: 'Lighthouse Labs',
+        dates: 'Aug 2019 - Mar 2021',
+        desc: 'Working with the Marketing team to plan and execute the building of Static ReactJs sites for marketing and promotional purposes. Working with the internal staff to maintain and upgrade internal systems in React, Ruby on Rails, etc.',
+        keywords: [
+          'Customer Service',
+          'Student Mentorship',
+          'Project Management',
+          'Marketing',
+          'Front-End Development (HTML5, CSS3, CSS Grid, Flexbox)',
+          'JavaScript',
+          'jQuery',
+          'Bootstrap',
+          'ReactJS',
+          'Ruby On Rails',
+          'JSX',
+          'Active Record'
+        ],
+        position: 'Front-End Developer',
+      },
+      {
+        company: 'FREELANCE WORK',
+        dates: 'AUG 2010 - PRESENT',
+        desc: 'Managing and completing websites for a variety of clients. Projects are handled from initial contact with the client to the final launch including Concepts, Wireframing, Design, Development, Migration, Content Management, Testing, and Launching. Supplying clients with Hosting and Host maintenance, including Server updates, and Hosting provider interaction when needed.',
+        keywords: [
+          'Business Management',
+          'Client Interaction',
+          'Client relations',
+          'Customer service',
+          'Project Management',
+          'Self Employment',
+          'Front-End Development (HTML5, CSS3)',
+          'Wordpress',
+          'PHP',
+          'MySQL',
+          'JavaScript',
+          'jQuery',
+          'Bootstrap',
+          'Photography',
+          'User Guide Creation'
+        ],
+        position: 'FULL STACK DEVELOPER'
+      },
+      {
+        company: 'FREECODECAMP',
+        dates: 'AUG 2017 - APR 2020',
+        desc: 'FreeCodeCamp is an online learning resource for Fullstack Development learning. FreeCodeCamp has broken learning Web Development into digestible chunks to allow people to learn how to code in their spare time. As a Global Community Manager, my biggest input into the FCC ecosystem is managing the moderators for the Facebook Groups and other social media.',
+        keywords: [
+          'Social Media Management',
+          'Facebook Group moderation',
+          'Community Organizing',
+          'Public speaking'
+        ],
+        position: 'GLOBAL COMMUNITY MANAGER'
+      },
+      {
+        company: 'MEDIAFACE, TORONTO',
+        dates: 'JUNE 2017 - OCT 2017',
+        desc: 'Building Responsive Worpdress Themes with Bootstrap. Tasks included building full websites, and Wordpress Plugins. As well as setting up SEO Tracking codes for Google Analytics, Facebook, and Bing.',
+        keywords: [
+          'Client Research',
+          'Client relations',
+          'Front-End Development (HTML5, CSS3)',
+          'Wordpress',
+          'PHP',
+          'MySQL',
+          'JavaScript',
+          'jQuery',
+          'Bootstrap',
+          'Technical SEO Implementation'
+        ],
+        position: 'FULLSTACK WEB DEVELOPER'
+      },
+      {
+        company: 'DIGITAL SHIFT MARKETING, TORONTO',
+        dates: 'FEB 2016 - OCT 2016',
+        desc: 'Building Responsive Worpdress Themes with Bootstrap. Tasks included building full websites to building landing pages for SEO A/B Testing. As well as setting up SEO Tracking codes for Google Analytics, Facebook, and Bing. Management duties included heading projects, as well as writing processes and procedures to be used by the development department and managing the IT needs of the company.',
+        keywords: [
+          'Development Team Management Training',
+          'Process writing',
+          'Client Research',
+          'Client relations',
+          'Customer service',
+          'Project Management',
+          'Front End Development (HTML5, CSS3)',
+          'Wordpress',
+          'PHP',
+          'MySQL',
+          'JavaScript',
+          'jQuery',
+          'Bootstrap',
+          'Technical SEO Implementation'
+        ],
+        position: 'FRONT-END DEVELOPER'
+      },
+      {
+        company: 'VENTUREWEB, TORONTO',
+        dates: 'SEPT 2013 – DEC 2014',
+        desc: `Working as a web consultant through VentureWeb for Accenture, as a site manager for Accenture's Careers Websites. My tasks included building full websites to building landing pages for SEO A/B Testing. As well as setting up SEO Tracking codes for Google Analytics, Facebook, and Bing.`,
+        keywords: [
+          'Development Team Management Training',
+          'Process writing',
+          'Client Research',
+          'Client relations',
+          'Customer service Project Management',
+          'Front-End Development (HTML5, CSS3)',
+          'Wordpress',
+          'PHP',
+          'MySQL',
+          'Javascript',
+          'jQuery',
+          'Bootstrap',
+          'Technical SEO Implementation'
+        ],
+        position: 'WEB CONSULTANT, FRONT END DEVELOPER'
+      },
+      {
+        company: 'NORTHLANDS, EDMONTON',
+        dates: 'SEPT 2012 – APR 2013',
+        desc: 'Primary responsibility for position was the updating and maintenance of the 13 web properties for Northlands. Acted as the primary contact for the hosting companies, and IT service providers that Northlands works with. This included managing, updating and maintaining the web servers, as well as working with the hosting companies to coordinate hardware maintenance and downtime. Additionally, responsible for internal system creation for managing new content to be added to the web properties.',
+        keywords: [
+          'Front-End Development (HTML5, CSS3)',
+          'JavaScript',
+          'jQuery',
+          'PHP',
+          'Wordpress',
+          'Drupal',
+          'Joomla!',
+          'User guide creation',
+          'Content Management',
+          'Web Platform Maintenance',
+          'Server Maintenance',
+          'Enterprise Level Company'
+        ],
+        position: 'SR. FRONT END DEVELOPER'
+      },
+    ]
+  }
+}
+
+export default resume