+++ src/globalAssets/data/skills.js
@@ -1,0 +1,132 @@
+/**
+ * Define data for the skills section.
+ */
+const skills = [
+  {
+    key: 'lang',
+    def: true,
+    title: 'Languages',
+    skillList: [
+      {
+        title: 'HTML5',
+        perc: '100%',
+      },
+      {
+        title: 'CSS3',
+        perc: '100%',
+      },
+      {
+        title: 'JavaScript',
+        perc: '80%',
+      },
+      {
+        title: 'TypeScript',
+        perc: '70%',
+      }, 
+      {
+        title: 'ES6',
+        perc: '100%',
+      },
+      {
+        title: 'Liquid',
+        perc: '50%',
+      }
+    ],
+  },
+  {
+    key: 'libAndFrame',
+    def: false,
+    title: 'Libraries and Frameworks',
+    skillList: [
+      {
+        title: 'NodeJS',
+        perc: '90%',
+      },
+      {
+        title: 'Express.Js',
+        perc: '95%',
+      },
+      {
+        title: 'ReactJS',
+        perc: '98%',
+      },
+      {
+        title: 'Webpack 5',
+        perc: '90%',
+      },
+      {
+        title: 'Babel',
+        perc: '90%',
+      },
+      {
+        title: 'ESLint',
+        perc: '100%',
+      },
+      {
+        title: 'Styled-Components',
+        perc: '100%',
+      },
+      {
+        title: 'React Testing Library',
+        perc: '80%',
+      },
+      {
+        title: 'Jest',
+        perc: '80%',
+      },
+      {
+        title: 'Cypress.io',
+        perc: '90%',
+      },
+      {
+        title: 'Semantic Ui',
+        perc: '90%',
+      },
+      {
+        title: 'Ajax',
+        perc: '100%',
+      },
+      {
+        title: 'Bootstrap',
+        perc: '100%',
+      },
+      {
+        title: 'JQuery',
+        perc: '100%',
+      },
+      {
+        title: 'Tailwind CSS',
+        perc: '50%',
+      },
+      {
+        title: 'Shopify',
+        perc: '50%',
+      },
+    ]
+  },
+  {
+    key: 'dbSys',
+    def: false,
+    title: 'Database Systems',
+    skillList: [
+      {
+        title: 'MySQL',
+        perc: '60%',
+      },
+      {
+        title: 'Postgressql',
+        perc: '60%',
+      },
+      {
+        title: 'MongoDb',
+        perc: '80%',
+      },
+      {
+        title: 'Restful Apis',
+        perc: '70%',
+      },
+    ]
+  },
+]
+
+export default skills