+++ src/types/data.interface.ts
@@ -1,0 +1,4 @@
+export interface skillItem {
+  perc: string,
+  title: string
+}